+++ server/src/tests/cases/compiler/include2.c
@@ -1,0 +1,8 @@
+// @files: includeFile.h
+
+test() {}
+
+// this include does not have a newline after it, which is legal
+// https://github.com/jlchmura/lpc-language-server/issues/184
+
+#include "includeFile.h"